feat: Properly fail with invalid host:port
