+++ lib/node-inspect.js
@@ -123,11 +123,13 @@ class NodeInspector {
       const attemptConnect = () => {
         ++connectionAttempts;
         this.stdout.write('.');
-        this.client.connect()
-          .then(null, () => {
+        return this.client.connect()
+          .then(() => {
+            this.stdout.write(' done');
+          }, () => {
             // If it's failed to connect 10 times then print failed message
             if (connectionAttempts >= 10) {
-              this.print(' failed to connect, please retry');
+              this.stdout.write(' failed to connect, please retry\n');
               process.exit(1);
             }
             return delay(500).then(() => attemptConnect());
